backend/src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
isodate==0.6.1
joblib==1.4.2
neo4j==5.20.0
networkx==3.3
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
scikit-surprise==1.1.4
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
